add katana

+++ setup.sh
@@ -10,16 +10,15 @@ set -euo pipefail
 
 
 ## install xcode
-#xcode-select --install
-touch /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
-PROD=$(softwareupdate -l |
-    grep "\*.*Command Line" |
-    head -n 1 |
-    awk -F"*" '{print $2}' |
-    sed -e 's/^ *//')
-softwareupdate -i "$PROD" --verbose
-rm /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
-
+if ! [[ "xcode-select -p" ]]; then
+    touch /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
+    PROD=$(
+        grep "Label:" |
+        awk -F ': ' 'END{print $2}'
+    )
+    softwareupdate -i "$PROD" --verbose
+    rm /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
+fi
 
 ## install Nix
 sh <(curl -L https://nixos.org/nix/install) --yes
